+++ src/windows/lib/vardlg.c
@@ -45,12 +45,12 @@ ADD_DWORD(unsigned char *p, DWORD dw)
 	return 4;
 }
 
-static int
+static size_t
 ADD_UNICODE_STRING(unsigned char *p, const char *s)
 {
 	WORD *w;
-	int i;
-	int len;
+	size_t i;
+	size_t len;
 
 	w = (WORD *)p;
 
@@ -64,7 +64,7 @@ ADD_UNICODE_STRING(unsigned char *p, const char *s)
 
 #define DWORD_ALIGN(p) { while ((DWORD)p % 4) *p++ = 0x00; }
 
-static int
+static size_t
 ADD_DLGTEMPLATE(unsigned char *dlg, short x, short y, short cx, short cy,
 		const char *caption, const char *fontname, WORD fontsize,
 		WORD n)
@@ -107,7 +107,7 @@ ADD_DLGTEMPLATE(unsigned char *dlg, short x, short y, short cx, short cy,
 	return (p - dlg);
 }
 
-static int
+static size_t
 ADD_DLGITEM(unsigned char *dlg, short x, short y, short cx, short cy,
 	    const char *label, WORD id, WORD type, DWORD style)
 {
@@ -276,7 +276,7 @@ vardlg_config(HWND hwnd, WORD width, const char *banner, WORD num_prompts,
 	hdc = GetDC(GetDlgItem(hwnd, cid)); /* assume one label is the same as all the others */
 
 	for (n = 0 ; n < num_prompts ; n++) {
-		GetTextExtentPoint32(hdc, prompts[n].prompt, strlen(prompts[n].prompt), &csize);
+		GetTextExtentPoint32(hdc, prompts[n].prompt, (int)strlen(prompts[n].prompt), &csize);
 		if (csize.cx > maxsize.cx)
 			maxsize.cx = csize.cx;
 		if (csize.cy > maxsize.cy)
@@ -352,7 +352,7 @@ vardlg_config(HWND hwnd, WORD width, const char *banner, WORD num_prompts,
 
 		hdc = GetDC(GetDlgItem(hwnd, id));
 
-		GetTextExtentExPoint(hdc, p, strlen(p), cx, &nFit,
+		GetTextExtentExPoint(hdc, p, (int)strlen(p), cx, &nFit,
 			pDx, &csize);
 
 		banner_y += csize.cy;
